add checks in linux scripts
Ignore-this: 662acfd7ac6df0e10733cc1afe6a26a0

darcs-hash:20110624121318-6831e-7e232e6ea0a2c328bfb85ae9044aeb9a6386b83d.gz

+++ test.sh
@@ -1,8 +1,21 @@
 #!/bin/bash
 
 runhaskell Setup configure --user --flags=testing
+if [ "$?" = "1" ]
+then
+    exit 1
+fi
 runhaskell Setup build
+if [ "$?" = "1" ]
+then
+    exit 1
+fi
 rm -f *.tix
 dist/build/skema-common-test/skema-common-test
+if [ "$?" = "1" ]
+then
+    exit 1
+fi
 hpc report *.tix
 hpc markup *.tix
+set -e
